+++ src/main/java/me/sharpjaws/sharpSK/hooks/mcMMO/ExprmcMMOPartyLevel.java
@@ -1,17 +1,15 @@
 package me.sharpjaws.sharpSK.hooks.mcMMO;
 
-import javax.annotation.Nullable;
-
-import org.bukkit.event.Event;
-
-import com.gmail.nossr50.party.PartyManager;
-
 import ch.njol.skript.classes.Changer;
 import ch.njol.skript.lang.Expression;
 import ch.njol.skript.lang.SkriptParser;
 import ch.njol.skript.lang.util.SimpleExpression;
 import ch.njol.util.Kleenean;
 import ch.njol.util.coll.CollectionUtils;
+import com.gmail.nossr50.party.PartyManager;
+import org.bukkit.event.Event;
+
+import javax.annotation.Nullable;
 
 public class ExprmcMMOPartyLevel extends SimpleExpression<Number> {
 	private Expression<String> s;
@@ -50,21 +48,21 @@ public class ExprmcMMOPartyLevel extends SimpleExpression<Number> {
 		if (mode == Changer.ChangeMode.SET) {
 			Number n = (Number) delta[0];
 			Integer n2 = n.intValue();
-			com.gmail.nossr50.party.PartyManager.getParty(s.getSingle(e)).setLevel(n2.intValue());
+			com.gmail.nossr50.party.PartyManager.getParty(s.getSingle(e)).setLevel(n2);
 		}
 		if (mode == Changer.ChangeMode.ADD) {
 			Number n = (Number) delta[0];
 			Integer n2 = n.intValue();
 			com.gmail.nossr50.party.PartyManager.getParty(s.getSingle(e))
-					.setLevel(PartyManager.getParty(s.getSingle(e)).getLevel() + n2.intValue());
+					.setLevel(PartyManager.getParty(s.getSingle(e)).getLevel() + n2);
 		}
 		if (mode == Changer.ChangeMode.REMOVE) {
 
 			Number n = (Number) delta[0];
-			Integer n2 = n.intValue();
-			if (n2.intValue() < PartyManager.getParty(s.getSingle(e)).getLevel()) {
+			int n2 = n.intValue();
+			if ((int) n2 < PartyManager.getParty(s.getSingle(e)).getLevel()) {
 				com.gmail.nossr50.party.PartyManager.getParty(s.getSingle(e))
-						.setLevel(PartyManager.getParty(s.getSingle(e)).getLevel() - n2.intValue());
+						.setLevel(PartyManager.getParty(s.getSingle(e)).getLevel() - (int) n2);
 			} else {
 				com.gmail.nossr50.party.PartyManager.getParty(s.getSingle(e)).setLevel(0);
 			}
